% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_chol}
\alias{Eigen_chol}
\title{Cholesky decomposition of a matrix}
\usage{
Eigen_chol(M)
}
\arguments{
\item{M}{a square symmetric positive-definite matrix, real or complex}
}
\value{
The upper triangular factor of the Cholesky decomposition of 
  \code{M}.
}
\description{
Cholesky decomposition of a real or complex matrix.
}
\details{
Symmetry is not checked; only the lower triangular part of 
  \code{M} is used.
}
\examples{
xx
}
