% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_cosh}
\alias{Eigen_cosh}
\title{Matrix hyperbolic cosine}
\usage{
Eigen_cosh(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The matrix hyperbolic cosine of \code{M}.
}
\description{
Matrix hyperbolic cosine of a real or complex square matrix.
}
