% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_det}
\alias{Eigen_det}
\title{Determinant of a matrix}
\usage{
Eigen_det(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The determinant of \code{M}.
}
\description{
Determinant of a matrix
}
