% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_exp}
\alias{Eigen_exp}
\title{Exponential of a matrix}
\usage{
Eigen_exp(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The exponential of \code{M}.
}
\description{
Exponential of a real or complex square matrix.
}
