% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_inverse}
\alias{Eigen_inverse}
\title{Inverse of a matrix}
\usage{
Eigen_inverse(M)
}
\arguments{
\item{M}{an invertible square matrix, real or complex}
}
\value{
The inverse matrix of \code{M}.
}
\description{
Inverse of a real or complex matrix.
}
