% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_isInvertible}
\alias{Eigen_isInvertible}
\title{Check invertibility}
\usage{
Eigen_isInvertible(M)
}
\arguments{
\item{M}{a matrix, real or complex}
}
\value{
A Boolean value indicating whether \code{M} is invertible.
}
\description{
Checks whether a matrix is invertible.
}
\examples{
set.seed(666L)
M <- matrix(rpois(25L, 1), 5L, 5L)
Eigen_isInvertible(M)
}
