% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_kernelDimension}
\alias{Eigen_kernelDimension}
\title{Dimension of kernel}
\usage{
Eigen_kernelDimension(M)
}
\arguments{
\item{M}{a matrix, real or complex}
}
\value{
An integer, the dimension of the kernel of \code{M}.
}
\description{
Dimension of the kernel of a matrix.
}
\examples{
set.seed(666L)
M <- matrix(rpois(35L, 1), 5L, 7L)
Eigen_kernelDimension(M)
}
\seealso{
\code{\link{Eigen_isInjective}}, \code{\link{Eigen_kernel}}.
}
