% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_log}
\alias{Eigen_log}
\title{Logarithm of a matrix}
\usage{
Eigen_log(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The logarithm of \code{M}.
}
\description{
Logarithm of a real or complex square matrix, when possible.
}
\details{
The logarithm of a matrix does not always exist. 
  See \href{https://eigen.tuxfamily.org/dox/unsupported/group__MatrixFunctions__Module.html#title7}{matrix logarithm}.
}
