% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_lsSolve}
\alias{Eigen_lsSolve}
\title{Linear least-squares problems}
\usage{
Eigen_lsSolve(A, b)
}
\arguments{
\item{A}{a \code{n*p} matrix, real or complex}

\item{b}{a vector of length \code{n}, real or complex}
}
\value{
The solution \code{x} of the least-squares problem \code{Ax = b} 
  (similar to \code{lm.fit(A, b)$coefficients}).
}
\description{
Solves a linear least-squares problem.
}
\examples{
set.seed(129)
n <- 7; p <- 2
A <- matrix(rnorm(n * p), n, p)
b <- rnorm(n)
lsfit <- Eigen_lsSolve(A, b)
b - A \%*\% lsfit # residuals
}
