% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_pow}
\alias{Eigen_pow}
\title{Matricial power}
\usage{
Eigen_pow(M, p)
}
\arguments{
\item{M}{a square matrix, real or complex}

\item{p}{a number, real or complex, the power exponent}
}
\value{
The matrix \code{M} raised at the power \code{p}.
}
\description{
Matricial power of a real or complex square matrix, when possible.
}
\details{
The power is defined with the help of the exponential and the 
  logarithm. See \href{https://eigen.tuxfamily.org/dox/unsupported/group__MatrixFunctions__Module.html#title8}{matrix power}.
}
