% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_rank}
\alias{Eigen_rank}
\title{Rank of a matrix}
\usage{
Eigen_rank(M)
}
\arguments{
\item{M}{a matrix, real or complex}
}
\value{
The rank of \code{M}.
}
\description{
Rank of a real or complex matrix.
}
