% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_sin}
\alias{Eigen_sin}
\title{Matrix sine}
\usage{
Eigen_sin(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The matrix sine of \code{M}.
}
\description{
Matrix sine of a real or complex square matrix.
}
