% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_sinh}
\alias{Eigen_sinh}
\title{Matrix hyperbolic sine}
\usage{
Eigen_sinh(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The matrix hyperbolic sine of \code{M}.
}
\description{
Matrix hyperbolic sine of a real or complex square matrix.
}
