\name{EigenR-package}
\alias{EigenR-package}
\alias{EigenR}
\docType{package}
\title{
\packageTitle{EigenR}
}
\description{
\packageDescription{EigenR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EigenR}
\packageIndices{EigenR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EigenR}

Maintainer: \packageMaintainer{EigenR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
