% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_cosh}
\alias{Eigen_cosh}
\title{Matrix hyperbolic cosine}
\usage{
Eigen_cosh(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The matrix hyperbolic cosine of \code{M}.
}
\description{
Matrix hyperbolic cosine of a real or complex square matrix.
}
\examples{
library(EigenR)
M <- toeplitz(c(1,2,3))
Eigen_cosh(M)
(Eigen_exp(M) + Eigen_exp(-M)) / 2 # identical
}
