% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_QR}
\alias{Eigen_QR}
\title{QR decomposition of a matrix}
\usage{
Eigen_QR(M)
}
\arguments{
\item{M}{a matrix, real or complex}
}
\value{
A list with the \code{Q} matrix and the \code{R} matrix.
}
\description{
QR decomposition of a real or complex matrix.
}
\examples{
M <- cbind(c(1,2,3), c(4,5,6))
x <- Eigen_QR(M)
x$Q \%*\% x$R
}
