% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_range}
\alias{Eigen_range}
\title{Range of a matrix}
\usage{
Eigen_range(M, method = "QR")
}
\arguments{
\item{M}{a matrix, real or complex}

\item{method}{one of \code{"LU"}, \code{"QR"}, or \code{"COD"}; the 
\code{"LU"} method is faster}
}
\value{
A basis of the range of \code{M}. With \code{method = "LU"}, the 
  basis is not orthonormal, while it is with \code{method = "QR"} and 
  \code{method = "COD"}.
}
\description{
Range (column-space, image, span) of a real or complex matrix.
}
