% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_range}
\alias{Eigen_range}
\title{Range of a matrix
Range (column-space, image, span) of a real or complex matrix.}
\usage{
Eigen_range(M, method = "LU")
}
\arguments{
\item{M}{a matrix, real or complex}

\item{method}{only \code{"LU"} is currently supported}
}
\value{
A basis of the range of \code{M}. With \code{method = "LU"}, the 
  basis is not orthonormal.
}
\description{
Range of a matrix
Range (column-space, image, span) of a real or complex matrix.
}
\examples{
xx
}
