% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenR.R
\name{Eigen_sinh}
\alias{Eigen_sinh}
\title{Matrix hyperbolic sine}
\usage{
Eigen_sinh(M)
}
\arguments{
\item{M}{a square matrix, real or complex}
}
\value{
The matrix hyperbolic sine of \code{M}.
}
\description{
Matrix hyperbolic sine of a real or complex square matrix.
}
\examples{
library(EigenR)
M <- toeplitz(c(1,2,3))
Eigen_sinh(M)
(Eigen_exp(M) - Eigen_exp(-M)) / 2  # identical
}
